Adding updated RentAWorld code from live code demonstration

+++ net-core/rent-a-world/RentAWorld.Models/Entities/Rental.cs
@@ -7,12 +7,12 @@ namespace RentAWorld.Models.Entities
         public int Id { get; set; }
         public string Title { get; set; }
         public string Description { get; set; }
-        public decimal AskingPrice { get; set; }
+        public int AskingPrice { get; set; }
         public bool Available { get; set; }
         public string Address { get; set; }
         public string City { get; set; }
 
-        // Metadata / Database specific properties
+        // Metadata
         public DateTime DateCreated { get; set; }
         public DateTime DateModified { get; set; }
         public string ModifiedBy { get; set; }
